Rename internal esplora modules, fix docs
